+++ model/Mascota/Mascota.py
@@ -1,0 +1,10 @@
+class Mascota:
+    def __init__(self, nombre, cedula_dueño, edad, especie, raza, caracteristicas, peso):
+        self.nombre = nombre
+        self.cedula_dueño = cedula_dueño
+        self.edad = edad
+        self.id = None 
+        self.especie = especie
+        self.raza = raza
+        self.caracteristicas = caracteristicas
+        self.peso = peso
