+++ model/Persona/Persona.py
@@ -1,0 +1,6 @@
+class Persona:
+    def __init__(self, cedula, nombre, edad, rol):
+        self.cedula = cedula
+        self.nombre = nombre
+        self.edad = edad
+        self.rol = rol
