+++ model/Veterinaria/Veterinaria.py
@@ -1,0 +1,7 @@
+class Veterinaria:
+    def __init__(self):
+        self.personas = [] 
+        self.mascotas = [] 
+        self.facturas = []
+
+
